pyproject.toml
[project]
name = "fyndora"
version = "0.1.0"
description = "specialized finance and compliance system designed to support decentralized fundraising networks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-allauth>=65.9.0",
    "django-environ>=0.12.0",
    "django-htmx>=1.23.0",
    "django-redis>=5.4.0",
    "faker>=37.3.0",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.11.11",
    "pytest-django>=4.11.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
]
prod = [
    "gunicorn>=23.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings_test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--tb=short",
    "--strict-markers",
    "--cov=apps",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]
markers = [
    "unit: Unit tests for models, utilities, and business logic",
    "integration: Integration tests for workflows and API endpoints", 
    "system: System tests for end-to-end functionality",
    "performance: Performance tests for load and stress testing",
    "slow: Slow running tests",
]
testpaths = ["tests"]
